requirements.txt
sc2reader
pyinstaller
Pillow